Expenses adapter updates

+++ app/src/main/java/com/example/android_resources/screens/budget/BudgetFragment.kt
@@ -38,7 +38,7 @@ class BudgetFragment : Fragment(), KoinComponent {
         return v
     }
 
-    fun getBalanceUntilDate(date: Date): Int {
+    fun getBalanceUntilDate(date: Date): Double {
         return presenter.getBalanceUntilDate(date)
     }
 }
